chore: upgrade rust from 1.63 to 1.67 (#51)

+++ src/writer/tests.rs
@@ -63,7 +63,7 @@ proptest! {
         let capacity_pages = memory.size();
         let min_pages_required = (offset + bytes.len() as u64 + WASM_PAGE_SIZE - 1) / WASM_PAGE_SIZE;
 
-        assert_eq!(capacity_pages, min_pages_required as u64);
+        assert_eq!(capacity_pages, min_pages_required);
     }
 
     #[test]
